% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawkeyline.R
\name{draw_key_line}
\alias{draw_key_line}
\title{ReSolve drawkeyline function. Internal.}
\usage{
draw_key_line(data, params, size)
}
\arguments{
\item{data}{Input}

\item{params}{Parameters}

\item{size}{Sizes}
}
\description{
Formats the legend of scatterplot function properly
}
