% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.cols.R
\name{expand.cols}
\alias{expand.cols}
\title{Data handler. Internal only.}
\usage{
expand.cols(dm, n)
}
\arguments{
\item{dat}{xts object or matrix}
}
\description{
Logical output function for minute data.
}
