% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.minute.data.R
\name{is.minute.data}
\alias{is.minute.data}
\title{ReSolve minute data detection}
\usage{
is.minute.data(dat)
}
\arguments{
\item{dat}{Input data source}
}
\description{
Logical output function for minute data
}
