% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.arguments.R
\name{ram.arguments}
\alias{ram.arguments}
\title{ReSolve Argument Check}
\usage{
ram.arguments(type = "base", user.args)
}
\arguments{
\item{type}{selection: 'base' and 'umbrella' are the options}

\item{user.args}{list: user argument list to check against for validity.}
}
\description{
Internal function to check argument validity across plotting functions.
}
