% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.bar.plot.R
\name{ram.bar.plot}
\alias{ram.bar.plot}
\title{ReSolve Bar Plotting Function}
\usage{
ram.bar.plot(dat,...)
}
\arguments{
\item{dat}{Data is assumed to be a matrix or vector of returns.}

\item{...}{Complete argument details are located in \code{\link[RAMplot:ram.arguments]{ram.arguments}}}
}
\description{
Base Plotting Functions: Bar and Waterfall
}
\examples{

## The function will call the bar plot by default.
## The waterfall plot is called as follows

ram.bar.plot(dat,emphasis=list(waterfall=T))

}
