% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.bootstrap.plot.R
\name{ram.bootstrap.plot}
\alias{ram.bootstrap.plot}
\title{ReSolve Bootstrap Plotting Function}
\usage{
ram.bootstrap.plot(dat,...)
}
\arguments{
\item{dat}{Data is assumed to be a matrix or vector of returns}

\item{...}{Complete argument details are located in \code{\link[RAMplot:ram.arguments]{ram.arguments}}}
}
\description{
Base Plotting Functions: Bootstrap
}
