% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.colors.R
\name{ram.colors}
\alias{ram.colors}
\title{ReSolve Color Function}
\usage{
ram.colors(n = 1, pcolor = "#00488d", lookup = NULL)
}
\arguments{
\item{n}{Number of colors to output. Max is 19.}

\item{pcolor}{Primary Color. Default is ReSolve "dark blue."}

\item{lookup}{Character input to lookup ReSolve specific color by chart name.}
}
\description{
Takes a numerical or lookup input and returns color values for R
}
