% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.ef.plot.R
\name{ram.ef.plot}
\alias{ram.ef.plot}
\title{ReSolve Efficient Frontier Plotting Function}
\usage{
ram.ef.plot(dat,...)
}
\arguments{
\item{dat}{Data is assumed to be a matrix or vector of prices}

\item{...}{Complete argument details are located in \code{\link[RAMplot:ram.arguments]{ram.arguments}}}
}
\description{
Base Plotting Functions: Correlation and Density
}
\examples{

ram.equity.plot(dat,...)

}
