% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.efmap.plot.R
\name{ram.efmap.plot}
\alias{ram.efmap.plot}
\title{ReSolve Efficient Frontier Map Plotting Function}
\usage{
ram.transition.plot(dat,...)
}
\arguments{
\item{dat}{Data is assumed to be a matrix or vector of rolling weights}

\item{...}{Complete argument details are located in \code{\link[RAMplot:ram.arguments]{ram.arguments}}}
}
\description{
Base Plotting Functions: Transition
}
