% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.equity.plot.R
\name{ram.equity.plot}
\alias{ram.equity.plot}
\title{ReSolve Equity Plotting Function}
\usage{
ram.equity.plot(dat,...)
}
\arguments{
\item{dat}{Data is assumed to be a matrix or vector of prices}

\item{...}{Complete argument details are located in \code{\link[RAMplot:ram.arguments]{ram.arguments}}}
}
\description{
Base Plotting Functions: Equity
}
