% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.pie.plot.R
\name{ram.pie.plot}
\alias{ram.pie.plot}
\title{ReSolve Bar Plotting Function}
\usage{
ram.pie.plot(dat, x.attributes = list(breaks = NULL, labs = NULL, labs.tilt =
  F), y.attributes = list(breaks = NULL, labs = NULL, show.values = F),
  titles = list(main = NULL, subtitle = NULL, legend.labels = NULL,
  legend.rows = NULL, caption = NULL, x = NULL, y = NULL, rounding = 1),
  emphasis = list(emph.column = NULL, hline = NULL, hline.size = NULL,
  hline.color = NULL, primary.column = 1, secondary.column = NULL, show.best.fit
  = T, waterfall = F))
}
\arguments{
\item{dat}{Input}

\item{x.attributes}{List: x axis modifiers}

\item{y.attributes}{List: y axis modifiers}

\item{titles}{List: plot titles}

\item{emphasis}{List: emphasis arguments}
}
\description{
Base plotting function.
}
