% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.plot.R
\name{ram.plot}
\alias{ram.plot}
\title{ReSolve Umbrella Plotting Function}
\usage{
ram.plot(dat, plot.type = "equity", start.date = NULL, end.date = NULL,
  x.attributes = list(show.day = F, breaks = NULL, labs = NULL, labs.tilt =
  F), y.attributes = list(start.val = 100, breaks = NULL, labs = NULL,
  trans.log = T, trans.percent = F, show.values = F, currency = "$"),
  titles = list(main = NULL, subtitle = NULL, legend.labels = NULL,
  legend.rows = NULL, caption = NULL, x = NULL, y = NULL, rounding = 1),
  emphasis = list(emph.column = NULL, hline = NULL, hline.size = NULL,
  hline.color = NULL, primary.column = 1, secondary.column = NULL, show.best.fit
  = T, waterfall = F, ring = F))
}
\arguments{
\item{dat}{Input}

\item{plot.type}{Selection: 'equity','scatter','bar','waterfall','correlation','transition'}

\item{start.date}{R formatted date input: Example 2000-12-31}

\item{end.date}{R formatted date input: Example 2000-12-31}

\item{x.attributes}{List: x axis modifiers}

\item{y.attributes}{List: y axis modifiers}

\item{titles}{List: plot titles}

\item{emphasis}{List: emphasis arguments}
}
\description{
Main plotting function.
}
