% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.preplot.R
\name{ram.preplot}
\alias{ram.preplot}
\title{ReSolve Plot Preparation Function}
\usage{
ram.preplot(dat, type)
}
\arguments{
\item{dat}{User input.}

\item{type}{The complete user list of preplot transforms:

\enumerate{
 \item{'standard'}{}
 \item{'melt'}{ # Corresponds to the density plot}
 \item{'waterfall'}{}
 \item{'correlation'}{}
 \item{'pie'}{}
 \item{'efmap'}{}
 \item{'ef'}{}
}}
}
\description{
Transforms input into a specific format for use with base plotting functions.
}
