% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.scatter.plot.R
\name{ram.scatter.plot}
\alias{ram.scatter.plot}
\title{ReSolve Scatter Plotting Function}
\usage{
ram.scatter.plot(dat,...)
}
\arguments{
\item{dat}{Data is assumed to be a matrix or vector of returns}

\item{...}{Complete argument details are located in \code{\link[RAMplot:ram.arguments]{ram.arguments}}}
}
\description{
Base Plotting Functions: Scatter
}
