% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.theme.R
\name{ram.theme}
\alias{ram.theme}
\title{ReSolve ggplot branding function}
\usage{
ram.theme(text.xaxis = 12, text.yaxis = 12, text.legend = 10)
}
\description{
Returns a ReSolve branded theme for ggplot2 objects.
}
