% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.minute.data.R
\name{is.minute.data}
\alias{is.minute.data}
\title{ReSolve minute data detection. Internal only.}
\usage{
is.minute.data(dat)
}
\arguments{
\item{dat}{xts object}
}
\description{
Logical output function for minute data.
}
