% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.arguments.R
\name{ram.arguments}
\alias{ram.arguments}
\title{ReSolve Argument Check}
\usage{
ram.arguments('type', ...)
}
\arguments{
\item{type}{selection: 'base' and 'umbrella' are the options}

\item{...}{The complete user list of arguments:

\describe{

 \item{\emph{x.attributes}}{
  \enumerate{
    \item{show.day = F}{}
    \item{breaks = NULL}{}
    \item{labs = NULL}{}
    \item{labs.tilt = F}{}
  }
 }

 \item{\emph{y.attributes}}{
  \enumerate{
    \item{start.val = 100}{}
    \item{breaks = NULL}{}
    \item{labs = NULL}{}
    \item{trans.log = T}{}
    \item{trans.percent = F}{}
    \item{show.values = F}{}
    \item{currency = '$'}{}
  }
 }

 \item{\emph{titles}}{
  \enumerate{
    \item{main = NULL}{}
    \item{subtitle = NULL}{}
    \item{legend.labels = NULL}{}
    \item{legend.rows = NULL}{}
    \item{text.legend = 10}{}
    \item{caption = NULL}{}
    \item{x = NULL}{}
    \item{y = NULL}{}
    \item{rounding = 1}{}
  }
 }

 \item{\emph{emphasis}}{
  \enumerate{
    \item{emph.column = NULL}{}
    \item{hline = NULL}{}
    \item{hline.size = NULL}{}
    \item{hline.color = NULL}{}
    \item{primary.column = 1}{}
    \item{secondary.column = NULL}{}
    \item{show.best.fit = NULL}{}
    \item{waterfall = F}{}
    \item{ring = F}{}
    \item{density = F}{}
    \item{ef.order = 'sharpe'}{}
  }
 }

 \item{\emph{ef}}{
  \enumerate{
    \item{custom.mean = NULL}{}
    \item{custom.covar = NULL}{}
    \item{plot.assets = NULL}{}
    \item{plot.CML = NULL}{}
    \item{asset.label.offset = 0.0025}{}
    \item{CML.label.offset = c(-0.01,0.06)}{}
    \item{asset.color = '#89d2ff'}{}
    \item{CML.color = '#9c3000'}{}
    \item{ef.color = '#00488d'}{}
    \item{rf = 0}{}
    \item{scale = NULL}{}
    \item{t.horizon = 10}{}
    \item{resampled = F}{}
    \item{n.samples = 1000}{}
    \item{lw = 0.5}{}
  }
 }

 \item{\emph{output}}{
  \enumerate{
    \item{file.name = NULL}{}
    \item{file.type = 'pdf'}{}
    \item{width = 10}{}
    \item{height = 5}{}
  }
 }

}}
}
\description{
This is the arguments module for the plotting functions. The user passes on arguments under the '.umbrella.' type which are then passed to the base plot function being called. This function also serves to guard against invalid argument parameters with a few basic user cases.
}
