% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.correlation.plot.R
\name{ram.correlation.plot}
\alias{ram.correlation.plot}
\title{ReSolve Correlation Plotting Function}
\usage{
ram.correlation.plot(dat,...)
}
\arguments{
\item{dat}{Data is assumed to be a matrix or vector of returns.}

\item{...}{Complete argument details are located in \code{\link[RAMplot:ram.arguments]{ram.arguments}}}
}
\description{
Base Plotting Functions: Correlation and Density
}
\examples{

# The function will call the correlation plot by default.
# The density plot is called as follows

ram.bar.plot(dat,emphasis=list(density=T))

}
