% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.dat.standard.R
\name{ram.dat.standard}
\alias{ram.dat.standard}
\title{ReSolve Data Standardizing Function}
\usage{
ram.dat.standard(dat)
}
\arguments{
\item{dat}{Input data source}
}
\description{
Takes an input and conforms it to RAM standards
}
