% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.pie.plot.R
\name{ram.pie.plot}
\alias{ram.pie.plot}
\title{ReSolve Pie Plotting Function}
\usage{
ram.pie.plot(dat,...)
}
\arguments{
\item{dat}{Data is assumed to be a matrix or vector of weights}

\item{...}{Complete argument details are located in \code{\link[RAMplot:ram.arguments]{ram.arguments}}}
}
\description{
Base Plotting Functions: Pie and Ring
}
\examples{

# The function will call the pie plot by default.
# The ring plot is called as follows

ram.bar.plot(dat,emphasis=list(ring=T))

}
