% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.plot.R
\name{ram.plot}
\alias{ram.plot}
\title{ReSolve Umbrella Plotting Function}
\usage{
ram.plot(dat,'plot.type',...)

See examples for argument use cases.
}
\arguments{
\item{dat}{Data input transformed for corresponding plot type.}

\item{plot.type}{Selection: 'equity','scatter','bar','waterfall','correlation','transition','pie','ring','density'}

\item{start.date}{R formatted date input: Example '2000-12-31'}

\item{end.date}{R formatted date input: Example '2000-12-31'}

\item{render}{Logical and TRUE by default. If FALSE the plot will not appear in R.}

\item{...}{Complete argument details are located in \code{\link[RAMplot:ram.arguments]{ram.arguments}}}
}
\description{
The umbrella plotting function for the RAMplot package. Data is assumed to be transformed appropriately for one of the following base plotting functions:
\describe{
 \item{\emph{:: equity}}{}
 \item{\emph{:: scatter}}{}
 \item{\emph{:: transition}}{}
 \item{\emph{:: waterfall}}{}
 \item{\emph{:: correlation}}{}
 \item{\emph{:: density}}{}
 \item{\emph{:: pie}}{}
 \item{\emph{:: ring}}{}
 \item{\emph{:: efmap}}{}
 \item{\emph{:: ef}}{}
}
}
\examples{

## Arguments can be called individually within a list.
ram.plot(dat,'equity',x.attributes=list(show.day=T))

}
