% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.preplot.R
\name{ram.preplot}
\alias{ram.preplot}
\title{ReSolve data preplot function}
\usage{
ram.preplot(dat, type = "standard")
}
\arguments{
\item{dat}{Input data source}

\item{type}{Selection: 'standard', 'waterfall', and 'correlation' are valid options}
}
\description{
Internal function which specifically formats an input for use base functions
}
